chore: api 스펙 변경 대응 - partyroom
