+++ src/features/partyroom/list-playback-histories/api/use-fetch-playback-histories.query.ts
@@ -1,9 +1,9 @@
 import { useQuery } from '@tanstack/react-query';
 import { AxiosError } from 'axios';
 import { QueryKeys } from '@/shared/api/http/query-keys';
-import { partyroomsService } from '@/shared/api/http/services';
+import { djsService } from '@/shared/api/http/services';
 import { APIError } from '@/shared/api/http/types/@shared';
-import { PlaybackHistoryItem } from '@/shared/api/http/types/partyrooms';
+import { PlaybackHistoryItem } from '@/shared/api/http/types/djs';
 import { useStores } from '@/shared/lib/store/stores.context';
 
 export default function useFetchPlaybackHistory() {
@@ -15,7 +15,7 @@ export default function useFetchPlaybackHistory() {
       if (!partyroomId) {
         throw new Error('partyroomId is not found. maybe you are not in the partyroom.');
       }
-      return partyroomsService.getPlaybackHistory({ partyroomId });
+      return djsService.getPlaybackHistories({ partyroomId });
     },
   });
 }
