+++ src/features/partyroom/list-playback-histories/index.ts
@@ -1,0 +1,1 @@
+export { default as PlaybackHistories } from './ui/list.component';
