--- src/features/partyroom/list-playback-history/index.ts
@@ -1,1 +1,0 @@
-export { default as PlaybackHistory } from './ui/list.component';
