+++ src/shared/api/http/types/djs.ts
@@ -14,6 +14,16 @@ export type UnregisterDjFromQueuePayload = {
   djId: number;
 };
 
+export type GetPlaybackHistoryPayload = {
+  partyroomId: number;
+};
+
+export type PlaybackHistoryItem = {
+  musicName: string;
+  nickname: string;
+  avatarIconUri: string;
+};
+
 export type SkipPlaybackPayload = {
   partyroomId: number;
 };
@@ -22,5 +32,6 @@ export interface DjsClient {
   registerMeToQueue: (payload: RegisterMeToQueuePayload) => Promise<void>;
   unregisterMeFromQueue: (payload: UnregisterMeFromQueuePayload) => Promise<void>;
   unregisterDjFromQueue: (payload: UnregisterDjFromQueuePayload) => Promise<void>;
+  getPlaybackHistories: (payload: GetPlaybackHistoryPayload) => Promise<PlaybackHistoryItem[]>;
   skipPlayback: (payload: SkipPlaybackPayload) => Promise<void>;
 }
