+++ src/widgets/partyroom-detail/ui/playback-history-panel.component.tsx
@@ -1,9 +1,9 @@
-import { PlaybackHistory } from '@/features/partyroom/list-playback-history';
+import { PlaybackHistories } from '@/features/partyroom/list-playback-histories';
 
 export default function PlaybackHistoryPanel() {
   return (
     <div className='mt-5 h-[500px] max-h-screen overflow-y-auto'>
-      <PlaybackHistory />
+      <PlaybackHistories />
     </div>
   );
 }
